requirements.txt
streamlit
python-dotenv
google-genai
google-cloud-aiplatform
